requirements.txt

requests
pandas
python-dotenv
tqdm
lxml
